Agregado casos donde el token esta expirado y arreglado boton de retroceso que mandaba al login y no borraba el token. Arreglado fragmento de tipos.

+++ Proyecto/app/src/main/java/com/proyecto/facilgimapp/model/dto/EntrenamientoDTO.java
@@ -15,7 +15,7 @@ public class EntrenamientoDTO implements Serializable {
     private int duracion;
 
     @SerializedName("tipoEntrenamientoId")
-    private Long tipoEntrenamientoId;
+    private Integer tipoEntrenamientoId;
 
     @SerializedName("usuarioId")
     private Integer usuarioId;
@@ -62,10 +62,10 @@ public class EntrenamientoDTO implements Serializable {
         this.duracion = duracion;
     }
 
-    public Long getTipoEntrenamientoId() {
+    public int getTipoEntrenamientoId() {
         return tipoEntrenamientoId;
     }
-    public void setTipoEntrenamientoId(Long tipoEntrenamientoId) {
+    public void setTipoEntrenamientoId(int tipoEntrenamientoId) {
         this.tipoEntrenamientoId = tipoEntrenamientoId;
     }
